targets split to separate files, makefile is now just a proxy for mage
