--- ci/commands/command_unittests.go
@@ -1,11 +1,0 @@
-package commands
-
-import (
-	"github.com/magefile/mage/sh"
-)
-
-// RunTests runs the tests
-func RunTests() error {
-	err := sh.RunV("go", "test", "-race", "-cover", "../...")
-	return err
-}
