+++ ci/util/cmd.go
@@ -1,0 +1,22 @@
+package util
+
+import (
+	"os"
+	"path"
+
+	"github.com/magefile/mage/sh"
+)
+
+// GetGoBinaryPath looks for the given binary in path, if not checks if it's in $GOPATH/bin
+func GetGoBinaryPath(binaryName string) (string, error) {
+	res, err := sh.Output("which", binaryName)
+	if err == nil {
+		return res, nil
+	}
+	gopath := os.Getenv("GOPATH")
+	binaryUnderGopath := path.Join(gopath, "bin", binaryName)
+	if _, err := os.Stat(binaryUnderGopath); os.IsNotExist(err) {
+		return "", err
+	}
+	return binaryUnderGopath, nil
+}
